cryze-bank/src/requirements.txt
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
pycryptodome==3.23.0
